+++ src/pdftl/cli/complete.py
@@ -30,6 +30,7 @@ HARDCODED_KEYWORDS = {
     "crop",
     "delete",
     "delete_annots",
+    "delete_attachments",
     "delete_bookmarks",
     "delete_blank",
     "delete_images",